Struct netfw_sys::INetFwPolicy2

#[repr(transparent)]pub struct INetFwPolicy2 { /* fields omitted */ }

Implementations

impl INetFwPolicy2

pub unsafe fn get_current_profile_types<'a, __0: Into<Param<'a, *mut c_long>>>(
&self,
profile_types_bitmask: __0
) -> HRESULT

pub unsafe fn get_firewall_enabled<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, *mut VARIANT_BOOL>>>(
&self,
profile_type: __0,
enabled: __1
) -> HRESULT

pub unsafe fn put_firewall_enabled<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, VARIANT_BOOL>>>(
&self,
profile_type: __0,
enabled: __1
) -> HRESULT

pub unsafe fn get_excluded_interfaces<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, *mut Variant>>>(
&self,
profile_type: __0,
interfaces: __1
) -> HRESULT

pub unsafe fn put_excluded_interfaces<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, Variant>>>(
&self,
profile_type: __0,
interfaces: __1
) -> HRESULT

pub unsafe fn get_block_all_inbound_traffic<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, *mut VARIANT_BOOL>>>(
&self,
profile_type: __0,
block: __1
) -> HRESULT

pub unsafe fn put_block_all_inbound_traffic<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, VARIANT_BOOL>>>(
&self,
profile_type: __0,
block: __1
) -> HRESULT

pub unsafe fn get_notifications_disabled<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, *mut VARIANT_BOOL>>>(
&self,
profile_type: __0,
disabled: __1
) -> HRESULT

pub unsafe fn put_notifications_disabled<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, VARIANT_BOOL>>>(
&self,
profile_type: __0,
disabled: __1
) -> HRESULT

pub unsafe fn get_unicast_responses_to_multicast_broadcast_disabled<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, *mut VARIANT_BOOL>>>(
&self,
profile_type: __0,
disabled: __1
) -> HRESULT

pub unsafe fn put_unicast_responses_to_multicast_broadcast_disabled<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, VARIANT_BOOL>>>(
&self,
profile_type: __0,
disabled: __1
) -> HRESULT

pub unsafe fn get_rules<'a, __0: Into<Param<'a, *mut INetFwRules>>>(
&self,
rules: __0
) -> HRESULT

pub unsafe fn get_service_restriction<'a, __0: Into<Param<'a, *mut INetFwServiceRestriction>>>(
&self,
service_restriction: __0
) -> HRESULT

pub unsafe fn enable_rule_group<'a, __0: Into<Param<'a, c_long>>, __1: Into<Param<'a, BSTR>>, __2: Into<Param<'a, VARIANT_BOOL>>>(
&self,
profile_types_bitmask: __0,
group: __1,
enable: __2
) -> HRESULT

pub unsafe fn is_rule_group_enabled<'a, __0: Into<Param<'a, c_long>>, __1: Into<Param<'a, BSTR>>, __2: Into<Param<'a, *mut VARIANT_BOOL>>>(
&self,
profile_types_bitmask: __0,
group: __1,
enable: __2
) -> HRESULT

pub unsafe fn restore_local_firewall_defaults(&self) -> HRESULT

pub unsafe fn get_default_inbound_action<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, *mut NET_FW_ACTION>>>(
&self,
profile_type: __0,
action: __1
) -> HRESULT

pub unsafe fn put_default_inbound_action<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, NET_FW_ACTION>>>(
&self,
profile_type: __0,
action: __1
) -> HRESULT

pub unsafe fn get_default_outbound_action<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, *mut NET_FW_ACTION>>>(
&self,
profile_type: __0,
action: __1
) -> HRESULT

pub unsafe fn put_default_outbound_action<'a, __0: Into<Param<'a, NET_FW_PROFILE_TYPE2>>, __1: Into<Param<'a, NET_FW_ACTION>>>(
&self,
profile_type: __0,
action: __1
) -> HRESULT

pub unsafe fn get_is_rule_group_currently_enabled<'a, __0: Into<Param<'a, BSTR>>, __1: Into<Param<'a, *mut VARIANT_BOOL>>>(
&self,
group: __0,
enabled: __1
) -> HRESULT

pub unsafe fn get_local_policy_modify_state<'a, __0: Into<Param<'a, *mut NET_FW_MODIFY_STATE>>>(
&self,
modify_state: __0
) -> HRESULT

pub fn get_interface<I: Interface>(&self) -> Option<I>

A safe version of QueryInterface.

If the backing class implements the interface I then a Some containing an ComPtr pointing to that interface will be returned otherwise None will be returned.